+++ simple_manga_downloader/modules/mangatown_com.py
@@ -18,7 +18,7 @@ class Mangatown(BaseManga):
         else:
             self.series_title = None
         self.manga_link = link
-        self.cover_url = None
+        self.covers = {}
         self.chapters = {}
 
     @request_exception_handler
@@ -41,7 +41,7 @@ class Mangatown(BaseManga):
 
         thumb = soup.find(class_="detail_info")
         if thumb:
-            self.cover_url = {self.series_title: thumb.img["src"]}
+            self.covers = {self.series_title: thumb.img["src"]}
 
         self.data = soup.find(class_="chapter_list").find_all("li")[::-1]
         return True